Cleaned up mod loading / example mods
I made a little class to serve as a container so mods feel a little cleaner.
Commented out some unnecessary debug logs

Mods must have a scene_mod = SceneMod(...) statement or the will fail to load.
If the mod scene has a loader set, it will replace the loader for the original scene.
If set, prefix and postfix will be added to the scene and executed when it is created.

This is a pretty good start for mod loading. As of yet you can technically insert a button which will call manager.change_scene() and insert your own custom scene, so one can add arbitrary content to the game...
or using the prefix/postfix methods one can add to the content of existing scenes, or modify their elements (respectively)... or overwrite the loader and stick in a scene of your own.
Very versatile.