More tweaks to the mod loader
I separated the before and after hooks inside of the dictionary so that things will happen in order rather than after decorators wrapping behavior that was supposed to happen before or what have you (depending on the load order of mods)

Note: I don't like having a get_scene method shoved into each scene, but that's the only way to keep the mod loading graceful, because if something like add_scene() is in the main part of the module, it'll run when it's imorted `meaning that the unmodified scene will be added to the list of scenes, then modified, but the modified version wouldn't be used`.

Note: defining the wrapper directly inside of the function loop resulted in stale references that ended up replacing scene loaders with each other very wrongly. Now there's a function to create the wrapper.