Draggable Entity Test
I copied the code for Ursina's draggable button prefab, but changed the parent class to Entity and used it in a mod to make a ghetto draggable kitty cat which will print to console when clicked (it's type and a few attributes it has) which scales with the scroll wheel.
I don't really get how the draggable class works in its totality at the moment, but I'll get a better grasp on that as I'm using it in a scene editor.

Added template for a mod which adds a mod menu (currently appears on the title menu as a german flag language button which takes you to the new empty scene inserted by the mod)